Add files via upload
The txt file shows questions that I came up with for data from the Sakila database which basically holds the information of a company that rents movies. 
The pdf file shows the results that I extracted by using SQL and I plotted the results in Excel. The presentation was made in Powerpoint.